+++ Assets/Models/Characters/CasualArt_Characters/Prefabs/NPC_2_Ann.prefab
@@ -238,9 +238,6 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 5029268dc4c94f249ab6b8d5e7181489, type: 3}
       insertIndex: -1
       addedObject: {fileID: 1412491912976854023}
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 5029268dc4c94f249ab6b8d5e7181489, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 2233463423920144782}
   m_SourcePrefab: {fileID: 100100000, guid: 5029268dc4c94f249ab6b8d5e7181489, type: 3}
 --- !u!1 &419489823695340539 stripped
 GameObject:
@@ -262,7 +259,7 @@ BoxCollider:
     serializedVersion: 2
     m_Bits: 0
   m_LayerOverridePriority: 0
-  m_IsTrigger: 1
+  m_IsTrigger: 0
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
@@ -339,33 +336,6 @@ NavMeshAgent:
   m_BaseOffset: 0
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
---- !u!54 &2233463423920144782
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 419489823695340539}
-  serializedVersion: 4
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_CenterOfMass: {x: 0, y: 0, z: 0}
-  m_InertiaTensor: {x: 1, y: 1, z: 1}
-  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ImplicitCom: 1
-  m_ImplicitTensor: 1
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 80
-  m_CollisionDetection: 0
 --- !u!4 &854995658654903236 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: -9021075015064496786, guid: 5029268dc4c94f249ab6b8d5e7181489, type: 3}
