+++ Assets/Models/Characters/CasualArt_Characters/Prefabs/NPC_3_Andrew.prefab
@@ -238,9 +238,6 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 4bbc55f6e5220004787a0ef8af97b4d8, type: 3}
       insertIndex: -1
       addedObject: {fileID: 628654916784894177}
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 4bbc55f6e5220004787a0ef8af97b4d8, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 2790955517342876317}
   m_SourcePrefab: {fileID: 100100000, guid: 4bbc55f6e5220004787a0ef8af97b4d8, type: 3}
 --- !u!1 &1597079517165192401 stripped
 GameObject:
@@ -262,7 +259,7 @@ BoxCollider:
     serializedVersion: 2
     m_Bits: 0
   m_LayerOverridePriority: 0
-  m_IsTrigger: 1
+  m_IsTrigger: 0
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
@@ -360,33 +357,6 @@ NavMeshAgent:
   m_BaseOffset: 0
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
---- !u!54 &2790955517342876317
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1597079517165192401}
-  serializedVersion: 4
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_CenterOfMass: {x: 0, y: 0, z: 0}
-  m_InertiaTensor: {x: 1, y: 1, z: 1}
-  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ImplicitCom: 1
-  m_ImplicitTensor: 1
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 80
-  m_CollisionDetection: 0
 --- !u!4 &1740125154929798382 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: -9021075015064496786, guid: 4bbc55f6e5220004787a0ef8af97b4d8, type: 3}
