+++ Assets/Models/Characters/CasualArt_Characters/Prefabs/NPC_4_Kevin.prefab
@@ -328,9 +328,6 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: f9a17a684cd09fa4c93c66c996cbeafc, type: 3}
       insertIndex: -1
       addedObject: {fileID: 2617328960918285654}
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: f9a17a684cd09fa4c93c66c996cbeafc, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: -3197762690006722907}
   m_SourcePrefab: {fileID: 100100000, guid: f9a17a684cd09fa4c93c66c996cbeafc, type: 3}
 --- !u!4 &5794608964374152856 stripped
 Transform:
@@ -357,7 +354,7 @@ BoxCollider:
     serializedVersion: 2
     m_Bits: 0
   m_LayerOverridePriority: 0
-  m_IsTrigger: 1
+  m_IsTrigger: 0
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
@@ -455,30 +452,3 @@ NavMeshAgent:
   m_BaseOffset: 0
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
---- !u!54 &-3197762690006722907
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6801926266836496039}
-  serializedVersion: 4
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_CenterOfMass: {x: 0, y: 0, z: 0}
-  m_InertiaTensor: {x: 1, y: 1, z: 1}
-  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ImplicitCom: 1
-  m_ImplicitTensor: 1
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 80
-  m_CollisionDetection: 0
