+++ Assets/Models/Characters/CasualArt_Characters/Prefabs/NPC_Teacher.prefab
@@ -232,9 +232,6 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 5e01548f1df0bc94187c4df7cfc1f61d, type: 3}
       insertIndex: -1
       addedObject: {fileID: 6483275671348123669}
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 5e01548f1df0bc94187c4df7cfc1f61d, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 355063512812460885}
   m_SourcePrefab: {fileID: 100100000, guid: 5e01548f1df0bc94187c4df7cfc1f61d, type: 3}
 --- !u!1 &5855638133929690808 stripped
 GameObject:
@@ -256,7 +253,7 @@ BoxCollider:
     serializedVersion: 2
     m_Bits: 0
   m_LayerOverridePriority: 0
-  m_IsTrigger: 1
+  m_IsTrigger: 0
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
@@ -293,33 +290,6 @@ MonoBehaviour:
   dooropenSound: {fileID: 0}
   narration: {fileID: 0}
   isLeftDoor: 1
---- !u!54 &355063512812460885
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5855638133929690808}
-  serializedVersion: 4
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_CenterOfMass: {x: 0, y: 0, z: 0}
-  m_InertiaTensor: {x: 1, y: 1, z: 1}
-  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ImplicitCom: 1
-  m_ImplicitTensor: 1
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 80
-  m_CollisionDetection: 0
 --- !u!4 &6867068159459634823 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: -9021075015064496786, guid: 5e01548f1df0bc94187c4df7cfc1f61d, type: 3}
