+++ Assets/Prefabs/Yutnori/YutnoriGame 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4a9f96ea1f3cefd4db647667b177a93f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
