+++ Assets/Scripts/Minigame/Yutnori/TotalYutnoriController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 14f6c16f80d89cd47a0d84d0b003bea3
+guid: 39903d2022e8468468479d13b5019e57
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
