--- Assets/Scripts/Minigame/Yutnori/YnrModeController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cd39e6361c5f1004b91ed7378adf2976
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
